ArtGalleryProjectHwarangDjangoBackend/requirements.txt
django>=3.2.14
djangorestframework>=3.12.4
djongo>=1.3.6
django-cors-headers>=3.11.0
pillow>=11.3.0